Display an indication of the score on the right hand side of the screen.  Proceed to the next level when the score maxes out.  Make each level harder than the last.  Lots of fixes to basic game mechanics.

+++ a2bejwld/types.h
@@ -35,7 +35,7 @@
 typedef uint8_t tSquare;
 typedef uint8_t tDirection;
 typedef uint8_t tGemType;
-typedef uint8_t tLevel;
+typedef uint16_t tLevel;
 typedef uint8_t tScore;
 
 typedef int8_t tPos;
